don't treat compilation warnings as errors
This is another place where the toolchain complains on my x86_64 Linux
system but not on my ARM Chromebook. Apparently Gtk 2.0's headers use
types that are defined in GLib 2.0 as being deprecated. I need to see
what accounts for that difference; I think maybe my Chromebook is using
an older Ubuntu.